LayoutTests/fast/viewport/viewport-83.html

<head>
    <title>Value can contain equals sign.</title>
    <meta name="viewport" content="width=height=830">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(160, 320, 480, 320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
