LayoutTests/fast/viewport/viewport-83-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "height" for key "width" not recognized. Content ignored.
viewport size 64x70 scale 5.000000 with limits [5.000000, 5.000000]
